refactor(ui): dock polish foundation — motion tokens, shared row kit
Prep layer for the dock refinement pass:
- Fix the dead content-morph fade (animate-fade-up ran 400ms and ignored the
  --tw-duration override) with a dedicated 200ms fade-in-fast token.
- Add a --color-glass-bot token; notch + cluster badge rings reference it
  instead of re-inlining the oklch literal, and the notch now matches the
  panel's bottom stop exactly.
- Relocate the LRow/LList/Tag row primitives into DockPanelKit (SinksPanel
  re-exports for back-compat) so Monitor/Incidents share one canonical row.
- DockPanel role dialog→region (non-modal, no focus trap expectation).
- Sinks health dot 6→8px; StatusChips onto the kit type scale; dock bar gets
  a max-width guard + drops status chips under the sm breakpoint.
