+++ apps/ui/src/index.css
@@ -97,6 +97,15 @@
   --animate-fade-up: fade-up var(--transition-duration-slow) var(--ease-emphasized) both;
   --animate-scale-in: scale-in var(--transition-duration-normal) var(--ease-emphasized) both;
   --animate-sheen: sheen 1.8s var(--ease-standard) infinite;
+  /* Fast opacity-only fade for the dock panel's content-morph (cluster swap):
+     a plain crossfade at the mockup's 200ms, distinct from the entrance
+     `fade-up` (which also translates and runs at the slow 400ms scale). */
+  --animate-fade-in-fast: fade-in-fast var(--transition-duration-fast) var(--ease-emphasized) both;
+
+  /* Bottom stop of the `surface-glass` gradient, promoted to a token so the
+     dock panel's notch and the cluster badge rings reference one value
+     instead of re-inlining the oklch literal. */
+  --color-glass-bot: oklch(0.2 0.009 255 / 0.82);
 }
 
 /* ── Polish layer: whisper-soft surface gradients ────────────────────────
@@ -140,6 +149,15 @@
   to {
     opacity: 1;
     transform: scale(1);
+  }
+}
+
+@keyframes fade-in-fast {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
   }
 }
